requirements.txt
# ── Core Deep Learning ─────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# ── Face Detection ─────────────────────────────────────────────────
facenet-pytorch>=2.5.3

# ── Computer Vision ────────────────────────────────────────────────
opencv-python>=4.8.0
Pillow>=10.0.0

# ── Data Science ───────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ── Visualization ──────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Web App ────────────────────────────────────────────────────────
streamlit>=1.28.0

# ── Utilities ──────────────────────────────────────────────────────
tqdm>=4.65.0
